requirements.txt
Django==2.1.3
djangorestframework==3.9.0
mistune==0.8.4
pytz==2018.7
dj-database-url==0.5.0
gunicorn==19.9.0
pytz==2018.7
whitenoise==4.1
psycopg2==2.7.5

